requirements.txt
functions-framework==3.*
yt-dlp
flask==2.*
requests
ffmpeg-python
groq
python-dotenv
podsearch
openai